// fe_inventarios_gei/src/app/models/usuario-minam.ts
export class UsuarioMinam {
  nIdUsuario: Number;
  sUsuario: String;
  sContracena: String;
  sNombres: String;
  sApellidos: String;
  sCorreo: String;
  sTelefono1: String;
  sEstadoRegistro: String;
  nTipo: String;
  nIdInstitucion: Number;
  sNombreInstitucion: String;
  nIdUser: Number;
  sEstado: String;
  sCambiaClave: String;
  nIdRol: any;
  sDniResp: String;
}